Change insert-object-hooks.pl to parse return type
Change insert-object-hooks.pl to detect the method's return type and pass that
return type into the generated macro. Also change the script to generate a call
to a different set of macros if the method is void. These changes pave the way
for adding reclamation table entry locking and unlocking in the swap preamble
macros.