Change MicroBenchmark worker thread to better measure overhead
The MicroBenchmark worker thread previously performed a lot of
method calls, in part to choose arrays and array elements randomly.
This commit changes the worker thread to choose arrays and array
elements deterministically and without any method calls, making it
easier to identify which bytecode instructions are executed during
the operation loop. This commit also changes the worker thread to
perform a fixed number of iterations every round, reducing the
number of timing method calls. Finally, this commit adds a loop
that performs integer operations for a configurable number of
iterations within the main operation loop, in order to measure the
overhead of Marvin for different workload mixtures.